Session redirect and Office side request response
If token is not set redirect to login

+++ backend/controller/users-controller.js
@@ -177,6 +177,11 @@ const resetPassword = async (req, res) => {
   res.status(204).json({ message: "Password updated!" });
 };
 const deleteUser = async (req, res, next) => {
+  const dataObject = JSON.parse(JSON.stringify(res.locals.user.userData));
+  if(dataObject.userType !== "Admin"){
+    console.log(dataObject);
+    return res.status(403).json({message: "No Access!"});
+  }
   await userModel
     .findByIdAndDelete(req.params.uID)
     .then(() => {
